Fix input type for multiplication chart creator to ensure float handling

+++ multiplication_chart_creator.py
@@ -1,4 +1,4 @@
-x=int(input("Enter a number to create its multiplication chart: "))
+x=float(input("Enter a number to create its multiplication chart: "))
 i=int(input("Enter the number of multiples to display: "))
 for i in range(1, i+1):
     print(i,"X",x,"=", i*x)
